Add error handling for missing --parsable flag for sbatch.
It's easy to forget to add the --parsable flag for sbatch when using a
custom script with --cluster-status. I just did it today. It took too
long to figure out the issue. Now the scripts emit relevant error
messages and return an exit code of 1, which causes Snakemake to stop
instantly. Before it just kept resubmitting jobs until it used up all
the restarts.